feat: Club unit tests done and clear

+++ app/Models/Services/ClubMemberService.php
@@ -2,11 +2,14 @@
 
 namespace App\Models\Services;
 
+use App\Exceptions\ModelDuplicateFoundException;
+use App\Exceptions\ModelNotFoundException;
 use App\Exceptions\NotFoundException;
 use App\Models\Club;
 use App\Models\ClubMember;
 use App\Models\Repositories\ClubRepository;
 use App\Models\Repositories\UserRepository;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Auth;
 
 class ClubMemberService
@@ -19,7 +22,7 @@ class ClubMemberService
         $this->userRepository = new UserRepository();
     }
 
-    public function addMembers(int $clubId, array $members): Club
+    public function addMembers(int $clubId, array $members): Model
     {
         $club = $this->clubRepository->getClub($clubId)->first();
         if (!$club) {
@@ -33,7 +36,7 @@ class ClubMemberService
             $newMember = new ClubMember();
             $newMember->club_id = $clubId;
             $newMember->user_id = $member;
-            $newMember->setMember();
+            $newMember->setMemberRole();
             $newMember->save();
         }
 
@@ -46,12 +49,29 @@ class ClubMemberService
         if (!$club) {
             throw new NotFoundException('Club not found');
         }
+
+        $isMember = ClubMember::where('club_id', $clubId)->where('user_id', Auth::user()->id)->first();
+        if ($isMember) {
+            throw new ModelNotFoundException('Už jsi členem klubu');
+        }
+
         $user = Auth::user();
 
         $newMember = new ClubMember();
         $newMember->club_id = $clubId;
         $newMember->user_id = $user->id;
         $newMember->role = 0;
+        $newMember->save();
+
+        return $newMember;
+    }
+
+    public function addOwner(int $clubId, int $userId): ClubMember
+    {
+        $newMember = new ClubMember();
+        $newMember->club_id = $clubId;
+        $newMember->user_id = $userId;
+        $newMember->setOwner();
         $newMember->save();
 
         return $newMember;
@@ -68,6 +88,10 @@ class ClubMemberService
         $member = ClubMember::where('club_id', $clubId)->where('user_id', $user->id)->first();
         if (!$member) {
             throw new NotFoundException('Member not found');
+        }
+
+        if ($member->isOwner()) {
+            throw new ModelDuplicateFoundException('Nemůžeš opustit klub, který jsi založil');
         }
 
         $member->delete();
@@ -152,7 +176,7 @@ class ClubMemberService
             throw new NotFoundException('Member not found');
         }
 
-        $member->setMember();
+        $member->setMemberRole();
         $member->save();
 
         return $member;
